test(server): split the capability-contract cases into test_meta_capabilities.py
test_meta_endpoint.py crossed the 600-line repo-shape cap once the
version-skew enforcement cases landed alongside main's desktop-updater
cadence cases. The capability-contract family (_cfg/_APP_COMPLETE and the
build_server_capabilities tests) is a self-contained unit that only needs
Settings and the pure builder, so it moves to a sibling file verbatim.

No test bodies changed and none were dropped; the allowlist does not grow.
